[tests] separate tests into online and offline
