+++ tests/linked/resources/file.txt
