+++ tests/linked/resources/file2.txt
