+++ tests/linked/resources/test_folder/sub_folder_2/sub_sub_file.txt
