+++ tests/linked/test_cli.py
